removing the reader_utils.py, and moving read_text() to reader.py
